backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
requests
python-multipart